created LevelUtils, moved methods from AITools
